Arrow funtion practice
Working on this problem was a good learning experience for me. At first, I took some time to understand exactly what I needed to do. It was tricky to handle the rule that said the greet function should give back "undefined" for certain kinds of input, but it helped me become better at solving problems.To solve the problem, I had to use special rules for different situations, like when the input wasn't a word or when it was an empty word. This made me more familiar with how JavaScript works and how to work with words in it.
I tested different ideas until I found the right solution. This whole process made me more confident in my coding skills and taught me a lot about how to handle different kinds of information in JavaScript. Overall, it made me a better developer.