Find-Bug
You have been given a code snippet in the index.js file. You can view the file in the editor.

Running this code gives an error, instead of logging the text, can you find out what's the bug with this code and fix it. You should try to add as minimal characters as possible to fix this.